refactor(providers): share connector exception types (CHAOS-2413)
