+++ src/dev_health_ops/connectors/exceptions.py
@@ -1,35 +1,17 @@
-"""
-Exception types for connector operations.
-"""
+from dev_health_ops.exceptions import (
+    APIException,
+    AuthenticationException,
+    ConnectorException,
+    NotFoundException,
+    PaginationException,
+    RateLimitException,
+)
 
-
-class ConnectorException(Exception):
-    """Base exception for all connector errors."""
-
-
-class RateLimitException(ConnectorException):
-    """Raised when API rate limit is exceeded."""
-
-    def __init__(
-        self,
-        message: str = "API rate limit exceeded",
-        retry_after_seconds: float | None = None,
-    ) -> None:
-        super().__init__(message)
-        self.retry_after_seconds = retry_after_seconds
-
-
-class AuthenticationException(ConnectorException):
-    """Raised when authentication fails."""
-
-
-class NotFoundException(ConnectorException):
-    """Raised when a resource is not found."""
-
-
-class PaginationException(ConnectorException):
-    """Raised when pagination fails."""
-
-
-class APIException(ConnectorException):
-    """Raised when API returns an error."""
+__all__ = [
+    "APIException",
+    "AuthenticationException",
+    "ConnectorException",
+    "NotFoundException",
+    "PaginationException",
+    "RateLimitException",
+]
